Add connect-mongo to store session data in our mongodb
Be sure to do a "npm install" after pulling this in.

By default express-session stores in memory the session data for
each user. It's not recommended for production, and more importantly
isn't compatible with passport.socket.io which we need. It also
means that user logins will survive nodejs restarts.

Now session data will be stored in the "session" collection in our
project2 mongodb as seen here:

$ mongo
MongoDB shell version: 3.2.8
connecting to: test
> use project2
switched to db project2
> show collections
sessions
users
> db.sessions.find().pretty()
{
        "_id" : "dp5RSMQxk0fjpw7ATA1BZ9bTzA4Xp8Hm",
        "session" : "{\"cookie\":{\"originalMaxAge\":null,\"expires\":null,\"httpOnly\":true,\"path\":\"/\"},\"flash\":{},\"passport\":{\"user\":\"58a3525728076d187ef21049\"}}",
        "expires" : ISODate("2017-03-08T07:45:09.633Z")
}
>

You'll notice that the passport object won't exist when the user is logged out.

To delete all session do:

> db.sessions.remove({})