Make sure assignment reqs are met on Registration Form
"First and Last Name Fields"

"The username should be one word that allows alphanumeric characters, as well as the – and _ characters. Anything else that is put into that box should be considered invalid and rejected."

"The password should be at least 6 characters, but no other rules should be imposed other than that. I recommend checking the criteria both on the browser side and the server side."